STM32F7 I2CAdapter compiles again following condensation.

+++ src/Digitabulum/CPLDDriver/SPIBusOp.h
@@ -59,11 +59,12 @@ class SPIOpCallback;
 */
 class SPIBusOp {
   public:
+    //TODO: This is the new mechanism: uint8_t  xfer_params[4];                      // The address transfer lengths, preamble, etc...
     uint8_t* buf            = NULL;               // Pointer to the data buffer for the transaction.
     SPIOpCallback* callback = NULL;               // Which class gets pinged when we've finished?
     XferState xfer_state = XferState::UNDEF;      // What state is this transfer in?
     BusOpcode  opcode    = BusOpcode::UNDEF;      // What is the particular operation being done?
-    uint16_t bus_addr    = 0x0000;                // The address that this operation is directed toward.
+    uint8_t  bus_addr    = 0x0000;                // The address that this operation is directed toward.
     int16_t  reg_idx     = -1;                    // Optional register index. Makes callbacks faster.
     uint8_t  buf_len        = 0;                  // How large is the above buffer?
